Rename io.h to avoid problems with Windows
