+++ deps/opende/src/internal_io.h
